+++ app/lib/widgets/questionnaire/questions/annotated_scale_question_widget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:studyu_core/core.dart';
 
-import 'custom_slider.dart';
+import '../custom_slider.dart';
 import 'question_widget.dart';
 
 class AnnotatedScaleQuestionWidget extends QuestionWidget {
