+++ app/lib/widgets/questionnaire/questions/boolean_question_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:studyu_core/core.dart';
 
-import '../selectable_button.dart';
+import '../../selectable_button.dart';
 import 'question_widget.dart';
 
 class BooleanQuestionWidget extends QuestionWidget {
